// src/core/enums/campaign.status.ts
export enum CampaignStatus {
  PUBLISHED = 'PUBLISHED',
  SCHEDULED = 'SCHEDULED',
  STOPPED = 'STOPPED',
  SEND_NOW = 'SEND_NOW',
  IN_PROGRESS = 'IN_PROGRESS',
  DRAFT = 'DRAFT',
  FAILED = 'FAILED',
}

export enum CampaignUpdateStatus {
  SEND_NOW = 'SEND_NOW',
  STOP = 'STOP',
  POSTPONE = 'POSTPONE',
  DRAFT = 'DRAFT',
}